support pre resolve host in myloader
